web/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.3.0
Flask==1.1.1
Flask-Cors==3.0.9
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==2.6.0
MarkupSafe==1.1.1
openapi-spec-validator==0.2.8
pymongo==3.9.0
python-dateutil==2.8.0
PyYAML==5.4
requests==2.22.0
six==1.12.0
swagger-ui-bundle==0.0.5
urllib3==1.25.8
Werkzeug==0.15.5
python-dotenv==0.19.1
pika==1.2.0